Decode image to uint8

+++ hanser/datasets/classification/general.py
@@ -8,7 +8,7 @@ from hhutil.io import read_lines
 
 
 def decode(example):
-    image = tf.cast(example['image'], tf.float32)
+    image = tf.cast(example['image'], tf.uint8)
     label = tf.cast(example['label'], dtype=tf.int32)
     return image, label
